Update and rename Play a song to Play song- karaoke
This mini karaoke project was written in python. Hardcode part of the lyrics of the song A thousand years by James Arthur and add the background music, the lyrics being played in the terminal together with the corresponding part of the song. I used python libraries such as time, sleep to ensure the playback of the lyrics in accordance with the song. I used the pygame library to play the sound of the song and the functions pygame.mixer.music.load to insert the song and pygame.mixer.music.play(start=5) to start from second 5 of the song.